Feat: se  termina de agregar la funcionalidad de la HU 7
1. se ajustan modelos
2. se  crean metodos en repositorio
3. se agrega un logger desde la carpeta helper
4. se ajustan la dockerizacion con variables de entorno
5. se agregan pruebas para el nuevo metodo

+++ MicroservicioSolicitudes/domain/model/src/main/java/co/com/mrcompany/model/sqs/QuotaData.java
@@ -14,6 +14,6 @@ public class QuotaData {
     private  Integer         number;
     private  BigInteger      amount;
     private  BigDecimal      monthQuota;
-    private  BigInteger      tax;
+    private  BigDecimal      tax;
     private  Double          interest;
 }
